requirements.txt
Django~=3.1.7
django-taggit~=1.3.0
Pillow~=8.2.0
social-auth-app-django~=4.0.0
ImageHash~=4.2.0
extcolors~=1.0.0
